% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as_smtrx}
\alias{as_smtrx}
\title{Cast to sociometrics}
\usage{
as_smtrx(x)
}
\arguments{
\item{x}{A data frame.}
}
\value{
The original object with corresponding class information attached.
}
\description{
Given a data frame, performs basic checks on availability of Timestamp data and
 other common sociometric columns and attaches the corresponding classes to the object. The
 \code{read_interaction, read_body, read_audio} functions attach the correct class format to
 the dataframe. However, performing other functions on might erase this information which
 then prevents the corresponding \code{qexpl()} or \code{anonymize} methods from working.
}
