% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{cluster_ids.smtrx}
\alias{cluster_ids.smtrx}
\title{Clusters consecutive ids}
\usage{
\method{cluster_ids}{smtrx}(df, clusterOffSet = 1)
}
\arguments{
\item{df}{Dataframe with Badge.ID column}

\item{clusterOffSet}{Numeric start of cluster ID number}
}
\value{
Modified dataframe with additional column "clusterIDSeq".
}
\description{
Cluster timestamped rows into clusters based on Badge.ID column sequence. Simplest case is to
create a cluster while badge.IDs of consecutive rows are the same. If Badge.ID from row n to
row n+1 changes, the n+1 row falls into new cluster.
}
