% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{cluster_transit.interact}
\alias{cluster_transit.interact}
\title{Cluster interactions}
\usage{
\method{cluster_transit}{interact}(df, clusterOffset = 0,
  by.col = "clusterTS", max.r = 10)
}
\arguments{
\item{df}{A data.frame of proximity data}

\item{clusterOffset}{Integer. Starting count for dyad clusters.}

\item{by.col}{String. The name of the column indicating existing clusters (or sessions) which
the algorithm will try to merge. If  \code{cluster.interaction.ts} has been run, this should be
"clusterTS"}

\item{max.r}{Integer Maxium recursion depth for clustering dyads.}
}
\value{
A data.frame with an additional column "clusterDyad"  running from clusterOffset + nr of clusters found.
}
\description{
The algorithm helps to detect a "meeting", i.e. co-presence of badges even though not all badges
detect all other badges.

Badges are grouped together based upon transitivness, i.e. a_b, b_c, c_d all belong to the same cluster,
since a and d are connected through b_c link. Usually \code{cluster.interaction.ts} has already been
called on the dataset since original interaction data do not overlap. Once several badges have been
grouped within a session, this method can look for badges that are shared between sessions
and hence allow to merge sessions to a larger one.
}
