% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rem.R
\name{dynam_event_ego}
\alias{dynam_event_ego}
\title{Format ego centered event list for DyNAM}
\usage{
dynam_event_ego(x, type = "MinMax", vprefix = NULL, vpostfix = NULL,
  fixtime = F, ids = NULL)
}
\arguments{
\item{x}{A data frame.}

\item{type}{String indicating the type of transformation required. "MinMax" specifies start and ending
of timestamps for each badge and day.}

\item{vprefix}{String prefix for labeling vertices. This is useful in relation to matrix where col
or row names might be just numbers while vertices are names.}

\item{vpostfix}{String postfix for labeling vertices.}

\item{fixtime}{Logical. Creates for each \code{ids} a time slot entry at the minimum time value - 5 min.
and sets it to \code{present=T}. This new time slot can be used for attaching covariate matrix whose actors need
to be present.}

\item{ids}{Vector of ids. Used together with \code{fixtime}. Indicates the ids for which a time slot
entry will be created. Should be all actors in the event network.}
}
\value{
A data frame containing time, node, replace columns.
}
\description{
Event sequences can be associated with ego-centered attributes in DyNAM. This is a
 utility function to retrieve the start end timestamps of actors to construct their presence/absence
 event list. Other types of event sequences are possible.
}
