% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rem.R
\name{dynam_nodes}
\alias{dynam_nodes}
\title{Format nodes for DyNAM}
\usage{
dynam_nodes(x, vprefix = NULL, vpostfix = NULL, present = FALSE)
}
\arguments{
\item{x}{Data frame of actors or nodes with arbitrary covariates}

\item{vprefix}{Node prefix to be pasted into label column}

\item{vpostfix}{Node postfix to be pasted into label column}

\item{present}{Logical. Indicates if nodes are present/absent at start of observation}
}
\value{
data frame ready to be used with DyNAM models.
}
\description{
Utility function to format data frame of nodes for use with DyNAM.
}
