% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{hash_source}
\alias{hash_source}
\title{Map data type abbreviations to source names}
\usage{
hash_source(abbr = NULL)
}
\arguments{
\item{abbr}{String of abbreviated data types.}
}
\value{
List with source (sheet) name and likely S3 class for parsing
}
\description{
Sociometric DataLab exports several different data sheets with often
 cumbersome names. The following function convert abbreviated shortcuts
 to excel data sheets into actual sheet names and suggests the S3 class which
 determines the corresponding parsing function.
}
\examples{

#produce a list of all currently available abbreviations
hash_source()

#name of data sheet for Infrared data
hash_shource("IR")

#name of data sheet for front microphone volume
hash_shource("VOL_F")

}
