% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mread}
\alias{mread}
\title{Multiple file reads}
\usage{
mread(readfn, files, pattern, ...)
}
\arguments{
\item{readfn}{Read function for sociometric files. @seealso \code{read_interaction},
\code{read_audio}, or \code{read_body}.}

\item{files}{List of absolute file paths to be read. Usually produced by \code{list.files}}

\item{pattern.}{Regular expression pattern for extracting part of the file name.
Extracted fragment will be stored in extra data frame column "ses_info".}
}
\value{
Tibble
}
\description{
Utility function for reading multiple files. Allows to extract part of a file name through the
provided regular expression. This allows to add additional information when reading several sociometric
export files and combining them into a single data frame. An additional column will be added
containing information about each original source file.
}
