% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mreverse.R
\name{mreverse}
\alias{mreverse}
\title{Match reversed values}
\usage{
mreverse(x, cols, into, sep, as_fct)
}
\arguments{
\item{x}{Either single- or multi column data frame. In case of data frame containing more than
two columns, \code{cols} needs to identify either a single column (containing splittable
strings) or two columns that hold the value pair.}

\item{cols}{Vector of column names (or index) if \code{x} has more than two columns.}

\item{into}{String. Name of additional column to store shared identifier.}

\item{sep}{Single character for spliting single column string values}
}
\value{
Original dataframe with additional column (specified by \code{into}) containing
 shared identifier between value pairs.
}
\description{
Eliminates directionality of value pairs. Largely a wrapper around
 \code{\link{grp_reverse_pairs}} accepting single, double or multi-column data frame
 inputs. Whereas \code{grp_reverse_pairs} generates list of unique reverse-value pairs,
 \code{mreverse} then assigns these identifiers to all entries of the given
 data frame.
}
\details{
A given list of directed dyads (e.g. "A-B", "B-A", "A-C", "C-A") is converted into
 undirected dyads such that "A-B" and "B-A" are both represented as "A-B".

 Accepts both, a single character column to be split (e.g. "Women-Men") or two
 columns such as "Head", "Tail".

 In case a single column is provided, needs to contain a string that can be split by
 a given separater \code{sep}.
}
\examples{
#single column
x <- data.frame(a=c("A-B", "B-A", "A-B", "C-D", "D-C", "Z-P", "Z-P"), stringsAsFactors = F)
mreverse(x)

#two columns
x <- data.frame(a=c(1,2,3,4,5), b=c(2,1,4,3,5), d=c("A-B", "B-A", "C-D", "D-C", "F-F"), stringsAsFactors = F)
mreverse(x, cols=c(1,2))

#or again just the single column with string value pair
mreverse(x, cols=c("d"))

}
\seealso{
\code{\link{grp_reverse_pairs}}
}
