% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse.ego}
\alias{parse.ego}
\title{Parse ego-based data}
\usage{
\method{parse}{ego}(raw_df, format = NULL, as_posixct = T, ts_col = 1, tz = NULL)
}
\arguments{
\item{raw_df}{Data frame as result of \code{read_*} functions. One column holds a character representation of timestamp values.}

\item{format}{Formatting string for timestamp. \code{format=NULL} dispatches \code{probe_ts_format}
 to assign a format automatically from a pre-established list (most common formats). \code{format=1|2}
 is the index position to the vector of formats; \code{format="\%d/\%m/\%y"} indicates specific
 format to be used in  \code{\link[base]{strptime}}.
\enumerate{
  \item \code{"\%m/\%d/\%y \%H:\%M:\%OS"}
  \item \code{"\%Y-\%m-\%d \%H:\%M:\%OS"}
}}

\item{as_posixct}{Logical. If resulting timestamp should be converted to POSIXct object.}

\item{ts_col}{Index or name of colum that holds timestamp data.}

\item{tz}{String. Specify timezone. Default \code{tz=NULL} will use \code{Syst.timezone()}}
}
\value{
Data frame of class c("ego") of pattern c("Timestamp", "Badge.ID", "Measure","Source")
}
\description{
Many data sheets contain ego based data, i.e. one bage produces
 a certain measurement (Volumne, BM Activity, ...) per timestamp.
}
