% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse}
\alias{parse}
\title{Parse sociometric data sheets}
\usage{
parse(raw_df, format = NULL, as_posixct = T, ts_col = 1)
}
\arguments{
\item{raw_df}{Data frame usually returned by file read operations.}

\item{format}{String representation of timestamp format}

\item{as_posixct}{Logical. Format timestamps as \code{POSIXct}}
}
\value{
Object.
}
\description{
Sociometric data sheets are often manufacturer dependent and come in different
 column configurations. The generic \code{parse} function provides an interface to write
 specific functions for each indiosyncractic data format and convert it into a tidy data.
 New implementations should create the corresponding S3 class instance and implement the
 specific parse function.
}
