% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse.smtrx}
\alias{parse.smtrx}
\title{Parse timestamp data.}
\usage{
\method{parse}{smtrx}(raw_df, format = NULL, as_posixct = T,
  ts_col = 1, tz = NULL)
}
\arguments{
\item{raw_df}{Data frame as result of \code{read_*} functions. One column holds a character representation of timestamp values.}

\item{format}{Formatting string for timestamp. \code{format=NULL} dispatches \code{probe_ts_format}
 to assign a format automatically from a pre-established list (most common formats). \code{format=1|2}
 is the index position to the vector of formats; \code{format="\%d/\%m/\%y"} indicates specific
 format to be used in  \code{\link[base]{strptime}}.
\enumerate{
  \item \code{"\%m/\%d/\%y \%H:\%M:\%OS"}
  \item \code{"\%Y-\%m-\%d \%H:\%M:\%OS"}
}}

\item{as_posixct}{Logical. If resulting timestamp should be converted to POSIXct object.}

\item{ts_col}{Index or name of colum that holds timestamp data.}

\item{tz}{String. Specify timezone. Default \code{tz=NULL} will use \code{Syst.timezone()}}
}
\value{
Original data frame with timestamp column converted to datetime object. Data frame
 has class "smtrx".
}
\description{
Converts string representation of timestamp value into date time object. Most
 exported sociometric measurements are timestamped. During the export (from Datalab) and import
 process (into R), the timestamp is stored as a string representation which needs to be converted
 into a date time object. The \code{col} parameter indicates the position of the column which
 holds the timestamp values, mostly the first column. This method is called by most \code{parse}
 functions to make sure the uniform naming and formatting of the timestamp column across all
 different sociometric data types. If the \code{raw_df} timestamp data is already a POSIXct object
 the data frame is returned directly.
}
