% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.interact}
\alias{plot.interact}
\title{Plot sociometric data}
\usage{
\method{plot}{interact}(df)
}
\arguments{
\item{df}{Data frame of class "interact", "smtrx"}
}
\value{
A graphics object
}
\description{
Plot interaction data
}
