% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qexpl.R
\name{qexpl}
\alias{qexpl}
\title{Quick visual explorations of sociometric datasets}
\usage{
qexpl(
  x,
  measure = NULL,
  title = "",
  size = 1,
  colour = NULL,
  alpha = 1,
  leg.pos = "none"
)
}
\arguments{
\item{x}{Data frame with sociometric data.}

\item{measure}{String. For dataframes of type \code{"ego"}, indicates the column name to be mapped
to y-axis.}

\item{title}{String. Title for the graphic}

\item{size}{Numeric. Width of geom_line or size of geom_point.}

\item{colour}{String. Name of the column to map on aesthetic colour.}

\item{alpha}{Numeric 0..1. 0 is completely transparent, 1 is completely obscure.}

\item{leg.pos}{String. In case colour mapping is used, indicates position of the legend, either
"bottom", "top", "left", "right"}
}
\value{
graphic object
}
\description{
A set of standard visualizations of interaction, audio, and body movement data.
}
\examples{

data("smtrx")

#print interaction data
qexpl(df.interact)

#add unidirectional edge column, named "Pairs"
df.interact <- mreverse(df.interact, into="Pairs")

qexpl(df.interact, colour="Pairs", title="Interaction pairs")

#print volume
qexpl(df.vol)

#print pitch
df.pitch \%>\%
 filter(ids=c("3119", "3184")) \%>\%
 qexpl(size=.2, title="Pitch", leg.pos="right")



}
