% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_audio}
\alias{read_audio}
\title{Import audio data}
\usage{
read_audio(file, type, ses_info = F, replv = F, delim = "\\t",
  format = NULL, tz = NULL, na.rm = F, cls = NULL, ...)
}
\arguments{
\item{file}{Path to source data file (xls, xlxs or csv).}

\item{type}{Indicates the type of audio data to be imported (see details for available
abbreviations).}

\item{ses_info}{Logical. Extract and store session info from file path if available.}

\item{replv}{Anonymization. Default \code{replv=FALSE} will leave original (badge) IDs in place. Set
to \code{replv=TRUE} will replace IDs with numbers starting from 1..n. Provide \code{replv=data.frame} with
values: First column holds original values, second column replacement values.}

\item{delim}{Single delimiter character for reading CSV data. Ignored for Excel files.}

\item{format}{Optional format for parsing timestamp data. If no format is specified two
pre-established timestamp formats are tried ut. See  \code{\link{parse.smtrx}} for details.}

\item{tz}{String. Default \code{tz=NULL} will use system timezone (\code{Sys.timezone()}) and
assign to timestamp. Useful for explicitly setting other than system timezone for timestamp
data.}

\item{na.rm}{Logical. Calls \code{na.omit} on the entire data frame after conversion to
tidy format.}

\item{cls}{Vector of class names. Default \code{cls=NULL} uses pre-defined sociometric
classes associated with the \code{type} abbreviation. However, class names can be specified
explicitly as well.}
}
\value{
Tibble with data in tidy format
}
\description{
Read audio CSV, XLS or XLSX files from Sociometrics. Original Excel sheets have
 often a nested column structure where 1 or more badges produce data over several subcolumns.
 This strucutre will be converted into a tidy data format.
}
\details{
Excel file reading is performed by \code{readxl::read_excel} function. Column type
 specification might be required at times with the \code{col_type} parameter, passed via the "..."

Volume, pitch and frequencies are avaible for the front- and back microphone which
 can be indicated by the "_F" or "_B" suffix on each abbreviation. If no suffix is included for
 both microphone data sheets (back + front) will be loaded. The following abbreviations are
 available for the \code{type} parameter:

 \itemize{
   \item{"VOL[_F|_B]" - Volume. "VOL" will load both front- and back microphone sheet. "VOL_F"
    only front and "VOL_B" only back microphone data. Volume levels range between 0 and 1.
    Values < 0.01 indicate not speaking, 0.01 - 0.02 speaking quitely, 0.03 - 0.08 speaking louder, and
    > 0.08 speaking loudly.}
   \item{"PITCH[_F|_B]" - Pitch. Depending on the DataLab export settings pitch measures are
    aggregated over a given time interval, starting from 1 - 60+ seconds. Typical male fundamental
    frequency ranges from 85 to 180 Hz; typical adult female from 165 to 255 Hz.}
   \item{"SP" - Speech profile (no front/back microphone option). The speech profile indicates for each badge
    Speaking, Overlap, Listening, Silent, Total Speaking and Total Silent duration. Values depend on the
    chosen time interval: if the speech profile has been exported over a period of 60s, Speaking (and all
    other) measures indicate fractions over the 60 second period, ranging from 0 to 60.
    If the speech profile has been exported over 1 second intervals, the columns indicate values that
    range from 0 to 1 second. \emph{Speaking} indicates the total time fraction a particular badge
    wearer was speaking; \emph{Overlap} the time fraction a person was speaking while someone else
    was speaking; \emph{Listening} the time fraction a particular badge was silent while someone else was speaking; \emph{Silent} the time fraction nobody
    was speaking. \emph{Total Speaking} is the sum of Speaking + Overlap per badge.}
   \item{"PAR" - Speech participation (no front/back microphone option). Logical. Indicates if a particular
    badge was speaking or not during the given time interval.}
   \item{"VOL_MIR[_F|_B]" - Volume mirroring. \emph{Similar} indicates the similarity between volume
    readings between two badges and ranges between 0 (no match) and 1 (perfect match) within the
    given time interval. \emph{Lag} indicates the time lag between matches. }
   \item{"VOL_CON[_F|_B]" - Volume consistency of each badge’s front audio amplitude, as measured in
    Activity (volume) (front). Consistency ranges from 0 to 1, where 1 indicates no changes in speech
    amplitude, and 0 indicates the maximum amount of variation in speech amplitude.}
   \item{"FRQ_[_F|_B]" - Dominant frequency. Contains three frequency bands hz_0, hz_1...hz_2 and
    corresponding amplitude readings amp_0, ...amp_2. Converted to tidy format, the resulting
    tibble contains the usual Timestamp, Badge.ID column followed by \emph{Band} column indicating
    one of the three bands \emph{Band_0, ... Band_2}} and two further columns \emph{Hz} and \emph{Amplitude}.
    There are potentially 4 frequency bands shown, hz_0 & amp_0 is the strongest PEAK in cepstrum,
    hz_1 & amp_1 is the second strongest PEAK, and so on. If there are fewer than k peaks in cepstrum,
    the hz_k and larger values are empty. E.g if there are only two peaks in cepstrum, hz_2 and hz_3 are empty and not exported.
   \item{"TT" - Turn taking sheet. \emph{Speaking Segment}: Any continuous, uninterrupted length of
    speech made by a single person. \emph{Turns}: Turns are speaking segments that occur after and
    within 10 seconds of, another speaking segment. By default a speech segment must be made within
    10 seconds after the previous one ended in order to be considered a turn. \emph{Self-turn}: A
    speaker starts speaking, pauses for greater than 0.5 seconds (but less than 10 seconds), and then
    resumes speaking. \emph{Successful interruptions}: Person A is talking. Peron B starts talking over
     A. If Person A talks for less than 5 out of the next 10 seconds, then Person B successfully
     interrupted Person A. \emph{Unsuccessful interruptions}: Person A is talking. Peron B starts talking
     over A. If Person A talks for more than 5 out of the next 10 seconds, then Person B successfully
     interrupted Person A. \emph{Pause}: A pause is a period of time within which there is no speaking.
     All pauses are between .5s and 10s.
   }
 }
}
\seealso{
\code{\link{read_body}} \code{\link{read_interaction}}
}
