% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_interaction}
\alias{read_interaction}
\title{Import interaction data}
\usage{
read_interaction(
  file,
  type,
  undirect = F,
  replv = F,
  delim = "\\t",
  format = NULL,
  tz = NULL,
  cls = NULL,
  ...
)
}
\arguments{
\item{file}{Path to source data file (xls, xlxs or csv).}

\item{type}{Indicates the type of interaction data to be imported (see details for available
abbreviations).}

\item{undirect}{Logical. Convert directed to undirected edge: \code{A->B | B->A = A<->B}}

\item{replv}{Anonymization. Default \code{replv=FALSE} will leave original (badge) IDs in place. Set
to \code{replv=TRUE} will replace IDs with numbers starting from 1..n. Provide \code{replv=data.frame} with
values: First column holds original values, second column replacement values.}

\item{delim}{Single delimiter character for reading CSV data. Ignored for Excel files.}

\item{format}{Optional format for parsing timestamp data. If no format is specified two
pre-established timestamp formats are tried ut. See  \code{\link{parse.smtrx}} for details.}

\item{tz}{String. Default \code{tz=NULL} will use system timezone (\code{Sys.timezone()}) and
assign to timestamp. Useful for explicitly setting other than system timezone for timestamp
data.}

\item{cls}{Vector of class names. Default \code{cls=NULL} uses pre-defined sociometric
classes associated with the \code{type} abbreviation. However, class names can be specified
explicitly as well.}
}
\value{
Tibble with data in tidy format
}
\description{
Read CSV, XLS or XLSX files from Sociometrics that contain Infrared-, Bluetooth- or both data types.
}
\details{
Excel file reading is performed by \code{readxl::read_excel} function. Column type
specification might be required at times with the \code{col_type} parameter, passed via the "..."

The following data sheets can be specified by setting the appropriate \code{type}
 parameter:

 \itemize{
   \item{"IR" - Infrared data}
   \item{"BT" - Bluetooth detects}
   \item{"IR_BT" - Both, Infrared and Bluetooth detects.}
   \item{"RIC" - Reads the interactions spell (or durations) between nodes. It contains "from", "to",
    columns and start/end timestamps combining both Infrared and Bluetooth readings.}
   \item{"NIC" - Reads the exported matrix of absolute interaction counts between all badges. Does
    not contain any timestamp data.}
 }

 All specified types will produce a data frame with five columns. Timestamp, Badge.ID, Other.ID,
 RSSI and Source. The data frame format is the same for IR, BT, IR_BT imports. The RSSI (Radio
 Signal Strength Indictor) holds a numeric value for Bluetooth signals (usually ranging from -40 to
 -90) indicating the relative proximity during BT signal detects. Smaller values indicate a weaker
 signal. Infrared detects do not have a measure of signal strength. IR detects are stored as NA values
 in the RSSI column.
}
\examples{

#load interaction including both BT and IR.
df1 <- read_interaction(file="data/datalab-full.xlsx", type="IR_BT")
df1

#load only Bluetooth data, replace Badge ids with numbers starting 1...n
df2 <- read_interaction(file="data/datalab-full.xlsx", type="BT", replv=T)
df2

#load only Infrared data, replace Badge ids with provided list
replv <- data.frame(mv=c(3182, 3189, 3183, 3178, 3184, 3081), rv=c(LETTERS[1:5]))
df3 <- read_interaction(file="data/datalab-full.xlsx", type="IR", replv=replv)
df3

#load interaction data, adding an undirected pair column
df4 <- read_interaction(file="data/datalab-full.xlsx", type="IR_BT", undirect=T)
df4


}
\seealso{
\code{\link{read_audio}}, \code{\link{read_body}}
}
