% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rem.R
\name{rem_edge_list}
\alias{rem_edge_list}
\title{Format timestamped edge list for REM}
\usage{
rem_edge_list(df, rm_sup = T, replv = T, use_seq = F)
}
\arguments{
\item{df}{The sociometric dataframe, containing at least the Timestamp, Badge.ID, Other.ID columns}

\item{rm_sup}{Logical. Remove duplicated values from proximity/face-to-face data frame where timestamps are identical}

\item{replv}{Logical. Make sure that badge ids in dataframe start at 1...ns}

\item{use_seq}{Logical. If Other.ID is missing, simulate Other.ID with Badge.ID of next row}
}
\value{
dataframe / edgelist with three columns only. Numeric Timestamp, head and tail vertex ids of
corresponding edges.
}
\description{
Convert a timestamped data frame into an edge list to be used with
relevent rem.dyad() which requires a three column matrix: timestamp, head, tail.
Timestamp data have to start near "0". Badge ids will be retrieved in ascending order and reset
to start from 1..n. It is important that the order of entries in covariate vectors retrieved
with \code{rem_covars} has the same ascending order.
}
