% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rr_edgelist}
\alias{rr_edgelist}
\title{Construct edge list from round-robin matrix}
\usage{
rr_edgelist(x, directed = T, to.undir = "weight")
}
\arguments{
\item{x}{round-robin matrix}

\item{directed}{logical. If set to \code{TRUE} will retain a directed edge list. If set to
\code{directed=F} the \code{weight} of each edge will be calculated according to \code{to.undir}.}

\item{to.undir}{string or number. Specifies the method for collapsing directed to undirected
edge list. Possible values are "min", "max", "mean", "recip", "weight", or a numeric value.
See details.}
}
\value{
tibble. In case an \code{directed=T}, a tibble with three columns: head, tail, edge, weight.
 In case \code{directed=F} a tibble with two columns only: edge, weight.
}
\description{
The \code{data(gedii_rr)} of package \code{gediismtrx} provides a matrix of
 round robin scores for advice seeking social affinity and psychological safety. This
 method provides utility functionalities to convert the round robin matrix into an
 edge list which retains original (weighted) scores, dichotomizes the matrix at certain
 weights, or seeks the min, max, or mean values of ties.
}
\details{
Round robin ratings are directional where person A rates person B while person B can
 of course rate person A differently. On some occasions, there directional weights (or scores)
 need to be collapsed, i.e. the directional ties are converted into undirectional ones based upon
 certain rules to be controlled by the two parameters \code{directed} and \code{to.undir}.

 The \code{to.undir} parameter takes the following values:

 \describe{
   \item{numeric value}{A numeric value will dichotomize the round robin matrix at the provided
    value and set all ties with a weight >= to.undir to 1, or 0 otherwise.}
   \item{"min"}{Will retain the minimum weight of two given ties between node pairs. If the weight
    of A->B = 2 and the weight of B->A = 4, then 2 will be retained for both ties.}
   \item{"max"}{Will retain the maximum weight of two given ties between node pairs. If the weight
    of A->B = 2 and the weight of B->A = 4, then 4 will be retained for both ties.}
   \item{"mean"}{Will retain the mean weight of two ties between node pairs. Note that NAs (i.e.
    missing ratings in the matrix) will first be replaced by the overall mean value of all ratings
    of the matrix. If the weight of A->B = 2 and the weight of B->A = 3, the mean value for
    the node pair A-B is (2+3)/2  2.5.}
   \item{"recip"}{Retains only ties which have a reciprocal weight, i.e. both involved parties
    gave the same rating to each other. If weight of A->B = 2 and B->A is 2, then 2 will be
    retained; otherwise the weight of the A-B pair will be replaced with 0}
   \item{"weight"}{Will retain the original scores. No transformation is applied.}
 }

At the same time, \code{directed} controls if the resulting edge list maintains directed edges
or undirected. In the latter case, \code{\link{mreverse}} is applied to obtain an undirected
list of edges with their corresponding scores. A warning is issued if a undirected edge list is
requested but no method for \code{to.undir} specified. Similar, if a method other than "weight"
is specified for \code{to.undir}, which implies to collapse the directed to an undirected edge list,
a warning will be issued that the directionality of the weights will not be maintained.
}
\examples{
#construct matrix
rrmat <- matrix(c(1,0,1,2,0,0,1,1,1,2,3,4,5,2,3,4), ncol=4)

#retain only reciprocal ties of same weight
rr_edgelist(rrmat, directed=F, to.undir="recip")

#retain min value of ties
rr_edgelist(rrmat, directed=F, to.undir="min")

#retain "max" values of ties
rr_edgelist(rrmat, directed=F, to.undir="max")

#dichotomize matrix at weight 3 or above
rr_edgelist(rrmat, directed=T, to.undir=3)

#retain "mean" values
rr_edgelist(rrmat, directed=F, to.undir="mean")

}
