% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{unique_ids}
\alias{unique_ids}
\title{Retrieve unique values across several columns}
\usage{
unique_ids(x, cols = c("Badge.ID", "Other.ID"), decreasing = NULL)
}
\arguments{
\item{x}{Dataframe, usually containing sociometric data}

\item{cols}{Names or indices of columns to gather unique values from}

\item{decreasing}{Logical. Order badge ids in decreasing or increasing order.
\code{decreasing=NULL} leaves order unaffected.}
}
\value{
Vector of unique values over given columns
}
\description{
Convenient way to retrieve all IDs within a given dataset over several columns. In
 the context of sociometric data, these are usually directed edge lists; the function provides
 a quick access to all badges (nodes) that form part of an interaction network.
}
\details{
Contrary to \code{dplyr::distinct} if several columns are provided, the function
 does not find unique combinations across columns. Unique values are found "globally" by first
 stacking all columns into a single vector before finding distinct values.
}
\examples{
x <- data.frame(a=c(1:10), b=(1:10))
unique_ids(x, cols=c("a","b"))

#compare to

unique(x)

}
\seealso{
\code{\link{unique_dates}}
}
