% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rem.R
\name{dynam_event_presence}
\alias{dynam_event_presence}
\title{Control presence of nodes for DyNAM}
\usage{
dynam_event_presence(bm, interact, vprefix = NULL, vpostfix = NULL,
  fixtime = F, ids = NULL)
}
\arguments{
\item{bm}{Data frame with sociometric body activity}

\item{interact}{Data frame with sociometric interaction data}

\item{vprefix}{String prefix for labeling vertices. This is useful in relation to matrix where col
or row names might be just numbers while vertices used in DyNAM are full names}

\item{vpostfix}{String postfix for labeling vertices.}

\item{fixtime}{Logical. Creates for each \code{ids} an artificial time slot entry at the minimum time
value minus 5 minutes and sets it to \code{present=T} and \code{present=F} 1 minute after. This new time slot can be used for attaching
covariate matrix whose actors need to be present when linking the Dynamic Network Actors with
change events.}

\item{ids}{Vector of ids. Used together with \code{fixtime}. Indicates the ids for which a artificial time slot
entry will be created. This should include all badge ids for given group.}
}
\value{
A data frame containing time, node, replace columns.
}
\description{
Event sequences can be associated with ego-centered attributes in DyNAM. This is a
 utility function to retrieve the start and end timestamps of actors to construct their presence/absence
 event list.
}
\details{
This is a utility function for preparing Sociometric (Badges) data for use with DyNAM models. In
 particular it helps controlling the presence and absence of badges. By indicating when a given badge
 was present (switched on) but not involved in interactions versus simply absent (badge turned off) a
 more precise estimate of DyNAM rates/choice should be possible. This is also interesting for data
 that spans several days. By indicating when people come to work in the morning and leave work at night
 the time people are not at work and therefore can't interact is taken into account.

 We use the body activity readings in order to determine at what point in time badges were switched on
 and off. The occurence of the first BT (or face-to-face) detects in not reliable in this sense since
 people can be at work but not interacting. However, in some cases badges body activity readings malfunction,
 i.e. there is no body activity reading although we have interaction detects. Therefore, the overall
 attendance of people is derived from the combination of the body activity dataset and interaction dataset
 for a given day. We construct a dataframe indicating the timestamp when people clock in (switch on) the
 badge in the morning and leave (switch off) later on (in the evening). In case the body activity data
 is missing, we take the first and last occurance of interaction for the given day as starting and
 ending timestamp.

 The parameters `ids` and `fixtime` should be used together. In order to be able to add events and
 nodes in the DyNAM model together, we need to make sure that all nodes exist at the start of the
 timespan used in the model. Therefore we take the earlierst timestamp of the dataset minus 5 minutes
 and artificially indicate that all badges specified with `ids` are present. After one minute, we set
 `presence=F` in order to start with the real presence/absence of badges. Otherwise, DyNAM linkEvents
 throws an error for events whose corresponding nodes are not present.
}
