% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_audio}
\alias{read_audio}
\title{Import audio data}
\usage{
read_audio(file, type = c("VOL", "PITCH", "SP", "PAR", "VOL_MIR", "VOL_CON",
  "FRQ", "TT"), ses_info = F, replv = F, delim = "\\t", format = NULL,
  na.rm = F, cls = NULL, ...)
}
\arguments{
\item{file}{Path to source data file (xls, xlxs or csv).}

\item{type}{Indicates the type of audio data to be imported (see details for available
abbreviations).}

\item{ses_info}{Logical. Extract and store session info from file path if available.}

\item{replv}{Anonymization. Default \code{replv=FALSE} will leave original (badge) IDs in place. Set
to \code{replv=TRUE} will replace IDs with numbers starting from 1..n. Provide \code{replv=data.frame} with
values: First column holds original values, second column replacement values.}

\item{delim}{Single delimiter character for reading CSV data. Ignored for Excel files.}

\item{format}{Format for parsing of timestamp data. Holds a series pre-established
timestamp formats. See  \code{\link{parse.smtrx}} for details. Reading excel sheets recognizes
date time objects. CSV files read character representations of dates which need the \code{format}
specified.}

\item{na.rm}{Logical. Should NA be removed?}

\item{cls}{Character. Vector of class names. Default \code{cls=NULL} using pre-defined sociometric
classes and subclass names. The class name indicates the corresponding parser to read data files
and convert them into tidy format.}
}
\value{
Tibble with data in tidy format
}
\description{
Read audio CSV, XLS or XLSX files from Sociometrics. Original Excel sheets have
 often a nested column structure where 1 or more badges produce data over several subcolumns.
 This strucutre will be converted into a tidy data format.
}
\details{
Volume, pitch and frequencies are avaible for the front- and back microphone which
 can be indicated by the "_F" or "_B" suffix on each abbreviation. If no suffix is included for
 both microphone data sheets (back + front) will be loaded. The following abbreviations are
 available for the \code{type} parameter:

 \itemize{
   \item{"VOL[_F|_B]" - Volume. "VOL" will load both front- and back microphone sheet. "VOL_F"
    only front and "VOL_B" only back microphone data. Volume levels range between 0 and 1.
    Values < 0.01 indicate not speaking, 0.01 - 0.02 speaking quitely, 0.03 - 0.08 speaking louder, and
    > 0.08 speaking loudly.}
   \item{"PITCH[_F|_B]" - Pitch. Depending on the DataLab export settings pitch measures are
    aggregated over a given time interval, starting from 1 - 60+ seconds. Typical male fundamental
    frequency ranges from 85 to 180 Hz; typical adult female from 165 to 255 Hz.}
   \item{"SP" - Speech profile (no front/back microphone option). The speech profile indicates for each badge
    Speaking, Overlap, Listening, Silent, Total Speaking and Total Silent duration. Values depend on the
    chosen time interval: if the speech profile has been exported over a period of 60s, Speaking (and all
    other) measures indicate fractions over the 60 second period, ranging from 0 to 60.
    If the speech profile has been exported over 1 second intervals, the columns indicate values that
    range from 0 to 1 second. \emph{Speaking} indicates the total time fraction a particular badge
    wearer was speaking; \emph{Overlap} the time fraction a person was speaking while someone else
    was speaking; \emph{Listening} the time fraction a particular badge was silent while someone else was speaking; \emph{Silent} the time fraction nobody
    was speaking. \emph{Total Speaking} is the sum of Speaking + Overlap per badge.}
   \item{"PAR" - Speech participation (no front/back microphone option). Logical. Indicates if a particular
    badge was speaking or not during the given time interval.}
   \item{"VOL_MIR[_F|_B]" - Volume mirroring. \emph{Similar} indicates the similarity between volume
    readings between two badges and ranges between 0 (no match) and 1 (perfect match) within the
    given time interval. \emph{Lag} indicates the time lag between matches. }
   \item{"VOL_CON[_F|_B]" - Volume consistency. }
   \item{"FRQ_[_F|_B]" - Frequency bands. Contains three frequency bands hz_0, hz_1...hz_2 and
    corresponding amplitude readings amp_0, ...amp_2. Converted to tidy format, the resulting
    tibble contains the usual Timestamp, Badge.ID column followed by \emph{Band} column indicating
    one of the three bands \emph{Band_0, ... Band_2}} and two further columns \emph{Hz} and \emph{Amplitude}
   \item{"TT" - Turn taking sheet. }
 }
}
\seealso{
\code{\link{read_body}} \code{\link{read_interaction}}
}
