% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_body}
\alias{read_body}
\title{Import accelerometer data}
\usage{
read_body(file, type = c("BM", "BM_RATE", "BM_CON", "BM_MIR", "POS",
  "POS_ACT", "POS_RATE", "POS_MIR"), undirect = F, replv = F,
  delim = "\\t", format = NULL, cls = NULL, ...)
}
\arguments{
\item{file}{Path to source data file (xls, xlxs or csv).}

\item{type}{Indicates the type of accelerometer data to be imported (see details for available
abbreviations).}

\item{undirect}{Logical. Convert directed to undirected edge: \code{A->B | B->A = A<->B}}

\item{replv}{Anonymization. Default \code{replv=FALSE} will leave original (badge) IDs in place. Set
to \code{replv=TRUE} will replace IDs with numbers starting from 1..n. Provide \code{replv=data.frame} with
values: First column holds original values, second column replacement values.}

\item{delim}{Single delimiter character for reading CSV data. Ignored for Excel files.}

\item{format}{Format for parsing of timestamp data. Holds a series pre-established
timestamp formats. See  \code{\link{parse.smtrx}} for details. Reading excel sheets recognizes
date time objects. CSV files read character representations of dates which need the \code{format}
specified.}

\item{cls}{Character. Vector of class names. Default \code{cls=NULL} using pre-defined sociometric
classes and subclass names. The class name indicates the corresponding parser to read data files
and convert them into tidy format.}
}
\value{
Tibble
}
\description{

}
\details{
The following accelerometer data sheets can be read by setting the \code{type} parameter
 to one of the following values:
 \itemize{
   \item{"BM" - Body movement activity. Values < 0.002 indicate very low activity; values between
    0.003 - 0.19 moderate amounts of activity and > 0.2 movements like walking.}
   \item{"BM_RATE" - Body movement rate}
   \item{"BM_CON" - Body movement consistency}
   \item{"BM_MIR" - Body movement mirroring}
   \item{"POS" - Posture. Left-right / front-back.}
   \item{"POS_ACT" - Posture activity}
   \item{"POS_RATE"- Posture rate}
   \item{"POS_MIR" - Posture mirroring}
}
}
