% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rem.R
\name{rem_covars}
\alias{rem_covars}
\title{Convert covariate factors to logical vector}
\usage{
rem_covars(df_covars, ids = NULL, covar_v = NULL, covar_m = NULL,
  covar_t = "attr", intercept = F, idcol = NULL, ...)
}
\arguments{
\item{df_covars}{Dataframe or matrix which contains covariate information. This can either be
a dataframe where columns specify sociodemographic variables such as age, gender, etc.
or it can be a matrix which specifies co-location.}

\item{ids}{Filter for specific ids in df_covars. If \code{ids=NULL}, all
entries of the \code{df_covars} will be retrieved. If \code{idcol=NULL}, assumes that ids match
row index; otherwise extra column with ids has to be specified in param \code{idcol}.}

\item{covar_v}{Vector of length two. First entry specifies name of attribute to retrieve.
This has to correspond to the column name in df_covars. Second entry specifies reference value set
to \code{TRUE}. If only first entry is provided original attribute values are
returned and not a logical vector.}

\item{covar_m}{Vector of length two. First entry specifies name of the attribute to
retrieve. Second entry specifies difference function, usually either "==" for binary values
like gender or "-" difference for numeric values. Returns n x n matrix of difference between
data attributes.}

\item{covar_t}{String. Either of c("attr" | "co.loc" | "rr"). Indicating type of covariate data.
Usually this is either a vector or a matrix of data attributes. However, there are further two
special types, namely the colocation matrix (covar_t="co.loc") and the round robin formats (covar_t="rr").
Default value is "attr".}

\item{intercept}{Logical. If TRUE, returns logical/matrix with all entries set to T.}

\item{idcol}{Name or index of the column which should be used to filter for \code{ids}. Default
is \code{NULL} meaning that provided \code{ids} correspond to the row index.}
}
\value{
either logical vector or data matrix for given attribute containing difference between two node values.

Matrix
}
\description{
Converts a factor (or character) vector into a logical vector (or matrix)
 where the reference category is set to \code{TRUE} and all other levels to \code{FALSE}. The
 resulting logical vector (or matrix) is used as covariate parameter in the \code{relevent::rem.dyad()}
 function.

 There are two modes: either retrieves a logical vector or a n x n difference/similarity matix of
 node attributes.
}
\details{
The order of the values for the covariate vector will be identical to the order in the
 df_covar dataframe. It is important that the IDs of the nodes has an ascending order since
 \code{relevent::rem.dyad()} expects its nodes to be ordered from 1..n in ascending order as well.
 When converting a sociometric data frame to an edge list to be used with relevent package, the
 default behavior takes care that original values are replaced in ascending order starting with
 1..n.
}
\examples{

#Retrieve a numeric vector of the column "Tenure" in dataframe df.attr for badge ids 2,5,7,9
tenure <- remcovars(df.attr, ids=c(2,5,7,9), covar_v=c("Tenure"))

#Retrieve a logical vector of "Role" Column, now with "Team leader" as reference category and set to TRUE
isLeader <-  remcovars(df.attr, ids=ub, covar_v=c("Role", "Team leader"))

#Retrieve a n x n matrix of Gender (where n depends on the length of badge IDs in "ub") of "Gender", indicating
#TRUE if two badges have the same "Gender" or FALSE otherwise.
sameSex <- remcovars(df.attr, ids=ub, covar_m=c("Gender", "=="))

#Retrieve the three round robin matrix for given badges, where NAs are replaced by column mean.
remcovars(rr_csx[[team]], ids=ub, na.replace="mean", covar_t="rr")

}
\seealso{
\code{rem_edge_list},
}
