% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{unique_dates}
\alias{unique_dates}
\title{Retrieve unique dates of dataset}
\usage{
unique_dates(x, ts_format = "\%Y-\%m-\%d")
}
\arguments{
\item{x}{Vector of timestamped data}

\item{ts_format}{String. Indicating the format of the specified date}
}
\value{
String of unique day-dates.
}
\description{
Convenient way to retrieve all unique dates - usually days - within data set.
 Retrieves the unique days ignoring hour, minute and sec information.
}
\seealso{
\code{\link{unique_ids}}
}
